refactor(app): extract the grid selection and editing wiring
Row nav, cell selection, opening the inline editor on double-click,
Enter/Tab commit, the footer's append-row and delete-row toggles, and
discard/commit of the pending write buffer for both panes move to
`wire::edit`. `repaint_edits` is declared inside that cluster and moves
with it.
