+++ app/src/wire/mod.rs
@@ -5,6 +5,7 @@
 //! and calls them in order; the handler bodies themselves are unchanged.
 
 pub(crate) mod conn_form;
+pub(crate) mod edit;
 pub(crate) mod grid;
 pub(crate) mod settings;
 pub(crate) mod update;
